journal: split into records/fold/docs/store/memory modules
Line-for-line moves along the existing section banners; public paths unchanged via
re-exports. journal.rs 3,657 -> mod.rs 163 + five focused files.
